packages/sie_sdk/pyproject.toml
[project]
name = "sie-sdk"
version = "0.1.7"
description = "Search Inference Engine - Python SDK"
requires-python = ">=3.12,<3.13"
license = { text = "Apache-2.0" }
dependencies = [
    # HTTP client (async)
    "aiohttp>=3.9,<4",
    # HTTP client (sync + shared helpers)
    "httpx>=0.28,<1",

    # WebSocket client
    "websockets>=14,<15",

    # Serialization
    "msgpack>=1.1,<2",
    "msgpack-numpy>=0.4,<1",

    # Data handling
    "numpy>=2,<3",
    "pillow>=11,<12",

    # Config parsing (bundle utilities)
    "pyyaml>=6,<7",
]

[project.optional-dependencies]
storage = [
    "boto3>=1.35,<2",
    "google-cloud-storage>=2.18,<3",
    "huggingface_hub[hf_transfer]>=0.26",
]
langchain = ["langchain-core>=0.3,<1"]
llamaindex = ["llama-index-core>=0.12,<1"]
haystack = ["haystack-ai>=2.7,<3"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/sie_sdk"]
